Move grammar files to Mongo-specific folder
