+++ src/mongo/services/completionItemProvider.ts
@@ -10,9 +10,9 @@ import { ParseTree } from 'antlr4ts/tree/ParseTree';
 import { ErrorNode } from 'antlr4ts/tree/ErrorNode';
 import { Token } from 'antlr4ts/Token';
 import { Db } from 'mongodb';
-import * as mongoParser from './../../grammar/mongoParser';
-import { mongoLexer } from './../../grammar/mongoLexer';
-import { MongoVisitor } from './../../grammar/visitors';
+import * as mongoParser from './../grammar/mongoParser';
+import { mongoLexer } from './../grammar/mongoLexer';
+import { MongoVisitor } from './../grammar/visitors';
 import { TextDocument, CompletionItem, Position, Range, CompletionItemKind } from 'vscode-languageserver';
 import SchemaService from './schemaService';
 import { LanguageService as JsonLanguageService } from 'vscode-json-languageservice';
