+++ src/mongo/services/mongoScript.ts
@@ -11,9 +11,9 @@ import { ErrorNode } from 'antlr4ts/tree/ErrorNode';
 import { Interval } from 'antlr4ts/misc/Interval';
 import { Token } from 'antlr4ts/Token';
 import { Db } from 'mongodb';
-import * as mongoParser from './../../grammar/mongoParser';
-import { mongoLexer } from './../../grammar/mongoLexer';
-import { MongoVisitor } from './../../grammar/visitors';
+import * as mongoParser from './../grammar/mongoParser';
+import { mongoLexer } from './../grammar/mongoLexer';
+import { MongoVisitor } from './../grammar/visitors';
 import { CompletionItemsVisitor } from './completionItemProvider';
 import SchemaService from './schemaService';
 import { LanguageService as JsonLanguageService } from 'vscode-json-languageservice';
